[ALPS07755471] apummu: buffer free flow
1. remove stable entry cnt and pay load (not record every mapping)
2. An API for DRAM buffer unmap
3. log level modify
4. Not pass session since MDW pass it in RV
5. is EVA uint64 for 34-bit addr
6. slb side band remove

<< Since entry cnt is removed, a RV patch must merge together >>

MTK-Commit-Id: a34a6b0308fc07f32226e1cffc119e7ad94755c2

CR-Id: ALPS07755471
Feature: apu_sys_firmware

+++ drivers/misc/mediatek/apusys/midware/2.0/mdw_mem.c
@@ -328,7 +328,7 @@ static int mdw_mem_map_create(struct mdw_fpriv *mpriv, struct mdw_mem *m)
 	struct mdw_mem_map *map = NULL;
 	struct scatterlist *sg = NULL;
 	int ret = 0, i = 0;
-	uint32_t eva = 0;
+	uint64_t eva = 0;
 
 	mutex_lock(&m->mtx);
 	get_dma_buf(m->dbuf);
